Modify tagging format in pipeline to include version
The tagging process in the GitHub Actions pipeline has been adjusted to include the version number. Previously, the tag name was only based on the date and time. Now, it also considers the version from the pom.xml file, which will help in creating more identifiable and descriptive tags for every pipeline run.